+++ compose.yml
@@ -53,7 +53,6 @@ services:
       - node_modules_ghost_offers:/home/ghost/ghost/offers/node_modules:delegated
       - node_modules_ghost_post-events:/home/ghost/ghost/post-events/node_modules:delegated
       - node_modules_ghost_post-revisions:/home/ghost/ghost/post-revisions/node_modules:delegated
-      - node_modules_ghost_posts-service:/home/ghost/ghost/posts-service/node_modules:delegated
       - node_modules_ghost_prometheus-metrics:/home/ghost/ghost/prometheus-metrics/node_modules:delegated
       - node_modules_ghost_security:/home/ghost/ghost/security/node_modules:delegated
       - node_modules_ghost_tiers:/home/ghost/ghost/tiers/node_modules:delegated
@@ -182,7 +181,6 @@ volumes:
   node_modules_ghost_offers: {}
   node_modules_ghost_post-events: {}
   node_modules_ghost_post-revisions: {}
-  node_modules_ghost_posts-service: {}
   node_modules_ghost_prometheus-metrics: {}
   node_modules_ghost_security: {}
   node_modules_ghost_tiers: {}
