--- ghost/posts-service/README.md
@@ -1,2 +1,0 @@
-# Posts Service
-
