--- ghost/posts-service/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-    PostsService: require('./lib/PostsService'),
-    PostsExporter: require('./lib/PostsExporter')
-};
